https://leetcode.com/explore/challenge/card/30-day-leetcoding-challenge/530/week-3/3305/
added support for doing this with post order.

I don't think doing this for in order is worth it, as I believe there could be more than one correct answer from in order.

E.g., [1 2 3] could be rooted at 1, 2 or 3.

So, I think, the BST structure is lost with an in order traversal.